Cope with daemons in private git.
The command as it was could not cope with a daemon coming from private git.

As I have it now, it will look in public git and conf directories for daemons
if they exist, and then in private if they don't. I believe this will work OK
for cities.writetothem, which has both a public and a private git checkout and
config directory - it's daemon bits will come from from the public one as the
did before.