make vhost config files group-readable so nagios plugins can read from them
At the moment, /data/vhost/*/conf/general is created mode 600.  The existing
monitoring scripts run as root so this isn't a problem for them, but nrpe
launches Nagios plugins as user 'nagios'.  Making the files mode 640 will
get around this, if the nagios user is a member of the vhost group.