use 'encoding' directive in vhosts.pl to set non-'UTF8' encoding for postgresql databases
Some databases, for historical reasons, use a non-default encoding.  When
restoring from backup, the encoding for the new database has to be the same,
or some data will fail to be ingested.  Ultimately all databases should use
the same encoding, but to make sure backups are useable we re-create databases
as-is.