Fix timeouts in run-with-lockfile so that errant commands are definitely killed
run-with-lockfile couldn't effectively kill timed-out processes because it always
got a shell to start them -- all that was being killed was the spawning shell.
Now we have an option '-e' which makes run-with-lockfile exec() the command directly.
Also added a further timeout in case a TERM isn't enough; after 10 seconds a KILL is sent.