Make sure repository is present for config update.
0c8dbd6 was too broad, in that on a first/timestamped deploy, the
repository was not present at the point it looked for *-example
files in order to update general/httpd/etc config.

Split create_or_update_files in two and nestle config update inside,
so the repository and settings file are created, then the config is
checked/updated, then the system files (web server, daemons) are
created if that passes.