Use temporary file for crontab.ugly
Deploying vhosts that have the crontab.ugly file in the repository
resulted in an empty crontab file in  /etc/cron.d being written,
effectively disabling cron for those vhosts.

I've not been able to determine the underlying cause of the bug - I
suspect it's something to do with the usage of pipes in the mugly()
call.

This commit switches back to using a temporary file for crontab.ugly
as a workaround.