Add a script for deploying Puppet environments
This script abstracts out the work done in the post-receive hook in the
Puppet repository to automatically deploy changes and new environments
and uses the contents of `script/` in the repository itself to update
and bootstrap the environment. (The hook itself will be updated in a
separate commit.)

The script can also be used in standalone mode to deploy into arbitrary
locations in the same way we deploy for production, so is useful for
development and testing. It also removes the assumption that the Puppet
master will be on the same host as the Git server.