[deploy-vhost] Refactor database configuration generation
Rather than generate the various formats for the database configuration
variables in the `deploy-vhost` script, factor this out into `Deploy.pm`
and the `write_settings_file` function.

`$conf->{database_configs}` is now an array of hashes containing each
database's configuration. `write_settings_file` uses this to generate
configuration in PHP, YAML and Rails formats as well as a set of
individual variables that can be used for other formats as needed.