Read STDIN in top-scope rather that inside foreach loop
Where a site has multiple vhosts with `push_actions` defined, only the
first encountered would be acted upon as `STDIN` was being read within
the scope of the foreach loop and the `@pushed_branches` array would
be empty on any further passes.

This moves the consumption of `STDIN` to the top-scope, preserving the state
of `@pushed_branches` across the subsequent loops.