Don't commit new dns and email files automatically
These files often need review and editing before they are ready for
deployment so we shouldn't assume that they are ready for committing.

This removes the git operations and prints out a message informing the
user what has been created and reminding them to check and commit before
deploying the changes.

Fixes #48.