Remove backends from balancers during deployments
This commit affects vhost deployments using the `--all` switch, aiming
to reduce the probability of downtime during deployments to multiple
backends.

The deploy process will now remove each node from the load-balancer pool
before updating the vhost. Once the deployment is complete, the script
will pause, then re-add the node to the load-balancers and check that it
is healthy on all load-balancers before continuing with the next node.

There is often a short period after a deploy where the back-end isn't
ready and the health checks sometimes take a few seconds to catch-up, so
this should prevent this from affecting end-users while we tune the
health checks.