Reverse usersync direction
The original usersync script assumed that it was being used to pull
updates from a reference host to another host. We're now using it in
the other direction, so want it to prefer local changes over remote
ones.

This will still allow changes on the remote node to propagate back up
to the local node (new users and groups and password changes) but will
overwrite any other changes on the remote rather than the local node.

A subsequent change may enforce the local files over the remote ones in
all cases, so that a reference host cane be used to guarantee the state
of remote hosts, rather than merging changes.