[debian-vhost-update] Remove `/dev/null` as overrides file argument
The overrides file argument to both `dpkg-scanpacakges` and
`dpkg-scansources` is optional. Having this here generates spurious
warnings that clutter up the output of this script.

(I suspect this is here because it is mentioned in the ancient and now
obsolete "Debian Repository HOWTO" as being required, which it may well
have been back in 2003.)