[debian-vhost-update] Permit multiple versions of a package
The `-m` switch will ensure all versions of a package are present in
the generated Packages metadata, rather then just the one detected as
the most recent.

We might as well do this and stop warnings being generated; we could
delete the old packages, but they are few and having older versions
may be useful.