[git hooks] Don't enumerate over all repos in our GitHub org when mirroring
Rather than enumerate over all the repositories in our organisation just
check for the actual repository. There shouldn't be more than one with
the same name in the mySociety namespace, so all we need to know is if
there is a repository and whether it is public or private.

This change is being made as calls to enumerate all repositories were
not reliable and even when this worked it was slow.