Add conf variable for $vcspath_install
This would be useful when processing ugly files.

For example we could add:
!!(*
  my $pro = `grep ^ENABLE_ALAVETELI_PRO $vhost_dir/$vcspath_install/config/general.yml | grep true | sed -n '\$='` || 0;
)!!

This would allow us to workout if Alaveteli Pro is enabled and later
conditionally add to the crontab if enabled. Currently we manually
maintain a list of vhosts in the config_crontab.ugly file which is
easily missed when upgrading a vhost.

We can't use the existing $vcspath variable as this is the final symlink
location of the vhost path. We need access to the timestamped path as
the ugly file processing happens before the symlinking.