[deploy-vhost] Experimental Docker integration
This adds initial support for Docker hosted services to the usual deploy
script. The behaviour of the script shouldn't change for non-Docker
vhosts.

It adds a custom stop and start command, skips Apache configuration and
restarts and ensures that there is a variable at `$conf->{docker}{stack}`
based on the vhost's FQDN with dots swapped out for underscores.