[deploy-vhost] Update how we pass back database password to Deploy.pm
This commit removes the nested call to `pgpw` in the value we pass back
to `Deploy.pm` in the database password field for writing out to a
vhost's `settings-autogen.pl` file. This doesn't add to the security of
this value (the file is only readable by the application user and the
same value is written directly into other configuration files) while
adding a layer of complexity.

It also adds an additional field, `password_escaped`, that has a version
of the password escaped as per RFC3986 that can be included safely in
database connection URIs. For now this is just added to the section of
individual `$db_config_*` variables to avoid having to add to all the
example configuration files for existing vhosts, none of which use this.