[update-vhosts-json] Don't open output file in binary mode
This script reads data in a text context then wrote it out in a binary
context, something Python 2 didn't care about but Python 3 does.

As we're dealing with JSON text here, this commit simply ensures that
the filehandle used for the output doesn't operate in a binary context
which I think should be fine in either case.