[deploy-vhost] Only flush Memcache and Varnish if found on app server
In the past we've always had Varnish and Memcache on every server. This
may not always be the case going forward, so we need to check before
blindly running each step of a `--flush`.

This commit checks for some relevant programs before attempting to
perform each flush. It's a tad naïve but should suffice for now as we'll
do a more thorough review in due course when we look at some wider
changes.

In the case of Varnish, if `varnishadm` isn't found the script will
suggest an alternative method.