[git-check-repositories] Update `get_expected_private_owner`
This updates and sorts alphanumerically the list of private repositories
with dedicate private groups.

In addition, it sets a default for `EXPECTED_OWNER` to `privatecvs`
which seems reasonable as this is the default private group - better to
check against this than not check at all, which is essentially what it
was doing before.