[mysociety] Connect to Varnish via TLS secured reverse proxy
Rather than connect directly to the Varnish management port on the load
balancers, this change will instead connect via a reverse proxy so as to
ensure the connection is encrypted on the wire.

Port mappings for the local end of the reverse proxy are expected to be
found in `/etc/mysociety/varnishadm` which will be populated by our
automation on management server instances from where it makes sense to
run with the `--all` flag.