[deploy-puppet-environment] Use CNAME alias when checking Puppet Server
When identifying the Puppet Server we can just use the CNAME alias
rather than doing another reverse lookup on the IP address.

None of these methods are completely reliable in all cases. If we're
using a floating IP at Brightbox and have set the CNAME to point to that
instead of the public server hostname to which it is currently mapped
then the subsequent match against the local hostname won't match. We'll
need to be careful to set the CNAME to the server's hostname for now.

Rather than pulling out the IP and then doing a reverse lookup on that