[new-git-repository] Remove IP address check for git.mysociety.org
We can't rely on the IP address for git.mysociety.org being bound to a
local interface on the server hosting it any longer.

For now this replaces the check with one that checks for the presence of
`/data/git`, which isn't perfect but is better than nothing.

This is a #minor update.

Fixes #69