[deploy-vhost] Remove legacy database schema checking code
We no longer use this method of checking database schemas are up to date
for a range of reasons, not least modern database migration management
being built-in to most frameworks and projects and broken assumptions
in some of the scripts used about database location and access.

This commit removes the calling code and the scripts themselves from the
working copy.

Closes #39