[Deploy.pm] Add `$cron_host` variable
This sets `$cron_host` to the first entry in `$conf->{servers}` so it
can be used in templates where a single system needs to be specified for
running certain cron tasks while others need to be everywhere. This is
intended to remove the need to hardcode server hostnames in conditional
statements in templates.