[mysociety] Check for `servers_internal` before removing from Varnish
If a server hostname is listed in both `servers` and `servers_internal`
then don't try to remove from the load-balancer, as it won't be in there
anyway.

This does rely on an updated `/data/vhosts.json` to work, so it adds a
call to `update-vhosts-json` at the top of the script.

See: https://github.com/mysociety/sysadmin/issues/1810