[deploy-vhost] Only stop/disable daemons when definitely present
Refactor `remove_daemons` to only attempt to disable a daemon where a
unit file is present as this might be called from `create_daemons` when
there are no daemons present.

Adds a call to stop the daemon to `remove_daemons` as on some paths
they may still be running. Stopping an already stopped daemon is fine.

This also consolidates the code a bit, without breaking the assumption
that there might be both a unit file and an init script present.

See: https://github.com/mysociety/misc-scripts/pull/80