[actions] Add an action that can build a Debian package
This adds an action that builds a Debian package, initially based on the
way we've done this in the packaging repository.

For the time being this needs to be triggered manually and the artefact
downloaded and added to our Debian repository, but we can add a step to
complete that subsequently.